pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "infraprobe"
version = "1.0.0"
description = "Infrastructure monitoring toolkit — probe your network, systems, and services"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [{name = "Talat Can Ayhan"}]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Networking :: Monitoring",
]
dependencies = [
    "typer>=0.12.0",
    "rich>=13.7.0",
    "pyyaml>=6.0",
    "pydantic>=2.7.0",
    "prometheus-client>=0.20.0",
    "requests>=2.31.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "black>=24.4",
    "flake8>=7.0",
    "mypy>=1.10",
    "bandit>=1.7",
    "pre-commit>=3.7",
    "types-PyYAML",
    "types-requests",
]

[project.scripts]
infraprobe = "infraprobe.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --strict-markers"
markers = [
    "slow: marks tests as slow",
    "integration: marks integration tests",
    "requires_root: tests requiring root privileges",
]

[tool.flake8]
max-line-length = 100
exclude = [".git", "__pycache__", "build", "dist", ".eggs"]
